// frontend/src/app/core/interfaces/change-password.interface.ts
// Generated by https://quicktype.io

export interface UpdatePasswordResponse {
  jwt: string;
  user: User;
}

export interface User {
  id: number;
  username: string;
  email: string;
  provider: string;
  confirmed: boolean;
  blocked: boolean;
  createdAt: string;
  updatedAt: string;
  birthday_date: string;
  name: string;
  lastName: string;
}

// Generated by https://quicktype.io

export interface UpdatePasswordBody {
  password: string;
  currentPassword: string;
  passwordConfirmation: string;
}